fix(vbc,types): Int.MIN const folds to i64::MIN (was 0)
Int.MIN resolved to 0 — every `Int.MIN < x` false, `from_int(Int.MIN)`
wrong, and ~10 primitives arithmetic-edge tests (*_at_min/*_neg_of_min).
Root cause (confirmed by trace: resolve_type_static_constant is never
called for Int.MIN — it is folded pre-codegen):
extract_const_literal_value (mod.rs) failed to fold the const body. For
`-9223372036854775808` = Neg(Lit(2^63)) it extracted 2^63 wrapped to
i64::MIN then `i64::MIN.checked_neg()` => None; the const then dropped to
body-compilation and resolved to 0. Fix: negate in i128 for a literal
operand so the most-negative literal folds to i64::MIN, and const-fold
Binary Add/Sub/Mul. primitives.vr MIN restored to the direct
`-9223372036854775808` (now folded correctly) — this also busts the
content-keyed stdlib-archive cache so the const re-bakes. Also fixed
integer_hierarchy IntegerKind::Int min/max from i128 (a stale
arbitrary-precision approximation) to i64 — Verum Int is 64-bit (BITS=64).

+++ crates/verum_types/src/integer_hierarchy.rs
@@ -175,7 +175,7 @@ impl IntegerKind {
     /// Get the minimum value for this integer type
     pub fn min_value(&self) -> i128 {
         match self {
-            IntegerKind::Int => i128::MIN, // Approximation for arbitrary precision
+            IntegerKind::Int => i64::MIN as i128, // Verum `Int` is 64-bit (BITS=64)
             IntegerKind::Int8 => i8::MIN as i128,
             IntegerKind::Int16 => i16::MIN as i128,
             IntegerKind::Int32 => i32::MIN as i128,
@@ -200,7 +200,7 @@ impl IntegerKind {
     /// Get the maximum value for this integer type
     pub fn max_value(&self) -> i128 {
         match self {
-            IntegerKind::Int => i128::MAX, // Approximation for arbitrary precision
+            IntegerKind::Int => i64::MAX as i128, // Verum `Int` is 64-bit (BITS=64)
             IntegerKind::Int8 => i8::MAX as i128,
             IntegerKind::Int16 => i16::MAX as i128,
             IntegerKind::Int32 => i32::MAX as i128,
